+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## master / unreleased
 
-## 1.12.0 in pgoress
+## 1.12.0 in progress
 
 * [CHANGE] Changed default for `-ingester.min-ready-duration` from 1 minute to 15 seconds. #4539
 * [CHANGE] query-frontend: Do not print anything in the logs of `query-frontend` if a in-progress query has been canceled (context canceled) to avoid spam. #4562